RF: added interfaces and aliases files
