+++ src/types/aliases.ts
@@ -1,0 +1,7 @@
+/**
+ * aliases.ts
+ * 
+ * This file contains any type aliases used
+ * These aliases definine complex or reusable types with a shorter, more descriptive name
+ * 
+ */
